rag-service/pyproject.toml
[project]
name = "dhara-rag-service"
version = "0.1.0"
description = "AI-powered legal research RAG pipeline for Bangladesh"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    "FlagEmbedding>=1.3.0",
    "google-generativeai>=0.8.0",
    "anthropic>=0.40.0",
    "openai>=1.58.0",
    "cohere>=5.13.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
